+++ src/games/commands/create-player.command.ts
@@ -1,0 +1,3 @@
+export class CreatePlayerCommand {
+  constructor(public readonly game: string, public readonly image: string) {}
+}